refactor(controller): DAO 직접 의존 제거, Service 레이어로 위임
