+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,37 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.service.ReservationService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/reservations")
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Reservation>> read() {
+        List<Reservation> reservations = reservationService.findAll();
+        return ResponseEntity.ok().body(reservations);
+    }
+
+    @PostMapping
+    public ResponseEntity<Reservation> create(@RequestBody ReservationRequest request) {
+        Reservation reservation = reservationService.save(request);
+        return ResponseEntity.ok().body(reservation);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationService.delete(id);
+        return ResponseEntity.ok().build();
+    }
+}
