+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,35 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.ReservationTime;
+import roomescape.service.ReservationTimeService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/times")
+public class ReservationTimeController {
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ReservationTime>> read() {
+        return ResponseEntity.ok().body(reservationTimeService.findAll());
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
+        ReservationTime savedTime = reservationTimeService.save(reservationTime);
+        return ResponseEntity.ok().body(savedTime);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationTimeService.delete(id);
+        return ResponseEntity.ok().build();
+    }
+}
